apps/fabric/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0
redis==5.0.1
aioredis==2.0.1
paho-mqtt==1.6.1
grpcio==1.59.3
grpcio-tools==1.59.3
protobuf==4.25.1
sqlalchemy==2.0.23
asyncpg==0.29.0
alembic==1.13.1
psycopg2-binary==2.9.9
opentelemetry-api==1.21.0
opentelemetry-sdk==1.21.0
opentelemetry-instrumentation-fastapi==0.42b0
opentelemetry-instrumentation-redis==0.42b0
prometheus-client==0.19.0
structlog==23.2.0
